Update ex_7b_csv_files.py
In this lesson will we will use the Python programming language to work with "comma-separated-value" or "CSV" files,  which are only one of several types of spreadsheet files.
Background: CSV files may include multiple data records with each data record located on its own line.  Each data record (i.e., each line in the file) may include multiple pieces of data separated by commas.  On each line, the pieces of data are always stored in the same sequence.  When saved, the file name for the data should end with the file extension ".csv"